+++ app/grails-app/domain/com/k_int/kbplus/IdentifierOccurrence.groovy
@@ -19,8 +19,6 @@ class IdentifierOccurrence {
             cre:   Creator
     ]
 
-
-
     static mapping = {
         id  column:'io_id'
         identifier column:'io_canonical_id'