+++ app/grails-app/domain/com/k_int/kbplus/Subscription.groovy
@@ -4,6 +4,7 @@ import com.k_int.kbplus.auth.*
 import com.k_int.properties.PropertyDefinitionGroup
 import com.k_int.properties.PropertyDefinitionGroupBinding
 import de.laser.helper.RDStore
+import de.laser.interfaces.DeleteFlag
 import de.laser.traits.AuditTrait
 import de.laser.domain.AbstractBaseDomain
 import de.laser.interfaces.Permissions
@@ -11,7 +12,9 @@ import de.laser.interfaces.TemplateSupport
 
 import javax.persistence.Transient
 
-class Subscription extends AbstractBaseDomain implements TemplateSupport, Permissions, AuditTrait {
+class Subscription
+        extends AbstractBaseDomain
+        implements TemplateSupport, DeleteFlag, Permissions, AuditTrait {
 
     // AuditTrait
     static auditable            = [ ignore: ['version', 'lastUpdated', 'pendingChanges'] ]
@@ -135,6 +138,11 @@ class Subscription extends AbstractBaseDomain implements TemplateSupport, Permis
         isPublic(nullable:true, blank:true)
         cancellationAllowances(nullable:true, blank:true)
         lastUpdated(nullable: true, blank: true)
+    }
+
+    @Override
+    boolean isDeleted() {
+        return RDStore.SUBSCRIPTION_DELETED.id == status?.id
     }
 
     // TODO: implement